Bugfix 1: DurableLog.tailReads were triggered based on ContainerMetadata.SequenceNumber. This was incorrect, since the sequence number would be updated prior to adding the operations to the log, which, in some cases, could trigger an assertion to fire (and fail the call erroneously). This is now fixed to use the last operation in the log as a trigger, which is the correct way of doing it.
Bugfix 2: Not checking for 'isClosed' when CacheManager unregisters a client. In some situations, when the CacheManager is closed before its clients, the clients would trip over this check when closing, thus having unwanted side effects (such as not closing properly). Since unregistering them simply means removing from a list so we don't have a reference to the object in memory, this check is relaxed and we are cleaning the client list upon closing the CacheManager.