Issue 637: (SegmentStore) Tier2 fencing using file chains (#923)
* Introduced SegmentHandle - a Storage abstraction that can hide (and cache) storage-specific 
information about the state of a segment.
Updated Storage and ReadOnlyStorage to make use of SegmentHandle.
Updated callsites and unit tests.

* Introduced the concept of Container Epoch. Every successful recovery will increase (and durably 
persist) the current Container Epoch. If two instances A and B of the same container have different 
epochs, the one with the higher epoch is the one that recovered most recently and the one with 
lower epoch should not be active anymore (because it was fenced out of Tier 1 DurableDataLog).


* Each SegmentContainer now has its own instance of a Storage adapter. This is in order to support 
container epochs, which is immutable throughout the lifetime of a Container Instance, and applies 
to all segments managed by that container. This is passed via the "initialize(epoch)" method of 
Storage, which is preferable to calling "openWrite(segment, epoch)" every time we need to open a 
segment.