Issue 959: (SegmentStore) Tier2 file compaction (#970)
1- ConcatOperation uses the native FileSystem.concat to concatenate all of the sources into the 
active file of the target Segment. This drastically simplifies the process (especially with recovery) 
and prevents us from potentially having a lot of files in case of segments with a high frequency of 
transactions. Due to restrictions in HDFS, this skips over empty source files, which are instead 
deleted.

2- OpenWriteOperation does the same thing after successfully fencing off a segment - it coalesces 
the previous files into one.

3- ReadOperation auto-refreshes the read handle if it detects an external file compaction happened.

4- Fixed a bug in StorageWriter where it could potentially register a SegmentAggregator but then fail to initialize it; as such the Writer would get stuck in an endless loop because it cannot use that aggregator at all. This was discovered because of #1 above, where HDFS was refusing to concat an empty file.