Issue 688: Docker compose (#1192)
This allows a user/developer to quickly spin up a pravega cluster using docker compose. This also fixes a few minor issues I encountered with the segmentstore's zkNamespace
and BookKeeper Ledger path.

Note that this does not yet use docker swarm, so it's only possible to use a single segmentstore and controller.

Partially fixes #688.