Bugfix for the following scenario: 1. A number of appends get written to Tier1 (up to offset T4) 2. Writer applies data to Tier2 up to offset (T3 < T4) 3. A truncation of data in Tier1 happens at offset T2 < T3 4. Container fails over to a new machine, but no full metadata checkpoint has been done after 3. At this point the last metadata checkpoint thinks that data in Tier2 is up to offset T0 < T1 (T0 recorded at the time of the previous checkpoint) 5. Immediately after recovery, a read request comes in. However, since during recovery we do not check for the latest state in Tier2 (since that would slow things down immensely), if this read requests data in (T0, T1), this data does not exist in Tier1 (hence in the cache) so the ReadIndex will return a "FutureRead" (until the data becomes available again). 6. It is possible that this returned FutureRead will never complete, because there may not be anything to trigger it (new appends, mergers or seals trigger it).
Previous fixes for this bug:
Attempt #1: When the writer starts processing the segment (asynchronously, after recovery), it already checks the state of the segment in Tier2 and updates the metadata accordingly. The attempt was to piggyback on this process and have it trigger those Future Reads. This fixed most of the problems, however there could have been a significant delay, or in case of other difficulties or races, it may never happen.
Attempt #2: Undo Attempt #1 and, instead of issuing a FutureRead, issue a StorageRead for any read that is not available in the cache but that has an offset less than the Segment's length. This fixed the perceived delays for Attempt #1, however it introduced races with appends and mergers, which first update the segment length in the metadata, and then populate the index (these are separate components and we cannot have a global sync lock for everything).

This (hopefully final) attempt does the following:
- Reverts Attempt #1 and Attempt #2.
- Introduces a new DurableLog operation called `StorageMetadataCheckpoint`. This is queued right before (and waited on) a truncation.
- This operation will collect the current Storage State (StorageOffset, StorageSealed, Deleted) for each segment and serialize it for later use
    - Upon recovery, when such an operation is encountered, the segments' metadata is updated accordingly.
    - This operation has a much smaller footprint than its sibling MetadataCheckpointOperation as it only deals with a few small fields
    - Those fields are not normally updated by regular operations, rather they're updated asynchronously by the Writer, which is why they cannot be recovered like the other fields.