Issue 1264: (SegmentStore) Fixed recovery sync to Tier2 (#1292)
* Introduces a new DurableLog operation called StorageMetadataCheckpoint. 
This is queued right before (and waited on) a truncation.

* This new operation collects the current Storage State (StorageOffset, StorageSealed,
Deleted) for each segment and serializes it for recovery purposes. Upon recovery,
when such an operation is encountered, the segment metadata is updated accordingly.

* This new operation has a much smaller footprint than its sibling 
MetadataCheckpointOperation as it only deals with a few small fields.

* Those fields are not normally updated by regular operations, and instead they're
updated asynchronously by the Writer, which is why they cannot be recovered like
the other fields.