Issue 1200: (SegmentStore) SegmentIds are sticky throughout Container lifetime (#1315)
SegmentIds (internal to the SegmentStore), once assigned, will have the same value throughout the life of a Segment. Instead of assigning a new SegmentId every time a Segment becomes active again, the SegmentId is persisted in the SegmentState (next to the attributes) and reloaded upon subsequent activations of the segment.