Issue 1428: Fix duplicate Segment Id bug (#1429)
* Fixed a bug in the ContainerMetadataUpdateTransaction that could potentially create duplicate 
Segment Ids. SegmentIds should be generated starting from the current Operation Sequence 
number (which guarantees a unique id to begin with), however due to a bug in the code, this was 
being generated off the newly assigned sequence number (which would only be set during 
recovery). This has been fixed to work as before.