Issue 1414: (SegmentStore) Tier1 Error Handling (#1439)
BookKeeperLog:
* Added ability to recover from BookKeeper ledger failures. In such cases, writes are retried, but not 
before opening a new ledger and updating the writes to be applied to the new ledger (with order 
being preserved).

* For writes with unknown outcome, we go to BK to fetch the LastAddConfirmed to see 
if the writes actually succeeded or not. This is done by open-fencing the ledger for which the write 
was applied, which guarantees agreement on the LastAddConfirmed value.

* Auto-closes when it detects a permanent write failure (and fails all subsequent queued writes). A 
permanent write failure is when a non-retryable error was detected or when the maximum number 
of retries for that write has been exhausted.

* All cancelled writes are now failed using CancellationException instead of the more obscure 
ObjectClosedException when the OperationProcessor is shutting down.

Segment Store (OperationProcessor et al):
* Auto-shuts down in case of DurableDataLog (Tier1) write failures - this is now done for any kind of 
failure, not just for DataCorruptionException or DataLogNotPrimaryException. The reason for this is 
that we do not have a clean way of completely ensuring write order in this case and the best way (at 
the moment) of reconciling this is to force a full container recovery. A bit of a brutal approach, but it 
avoids having to write complex code in order to support other, less drastic, approaches.

* All cancelled operations are now failed using CancellationException instead of the more obscure 
ObjectClosedException when the OperationProcessor is shutting down.

* DataFrameBuilder: no longer auto-flushing upon close. Since this object is owned throughout the lifetime of an OperationProcessor, it is closed whenever the OperationProcessor is closed, so no need for this feature anymore.