Issue 1034: EventStreamWriter uses an internal threadpool to handle retries. (#1356)
Event stream writer uses an internal thread-pool to do retries with this commit. One scenario where 
the client performs retries is when the client receives a sealedSegment from the segment store. In 
this scenario the client:

* Checks with the controller on the the next segment info;
* Refreshes / creates SegmentOutputStream for the current active segments;
* Resends unAck'ed events to the new SegmentOutputStream.

Using the internal thread pool also enables us to remove the AckFuture used in the eventWriter 
APIs.