Issue 1536: (SegmentStore) Container de-registration sporadic issues. (#1561)
* Fixes a bug in StreamSegmentContainerRegistry where it could erroneously fail a container 
registration request if that container has just been shut down. This was due to the fact that the 
internal de-registration happens in a callback which may be executed after the Future returned 
externally does. The solution provided handles all three possible cases: when a container fails 
during startup, when it shuts down (and results in failure or no failure) or when it fails while running.

* Fixes a bug in ServiceHelpers where it could be possible that callbacks passed to onStop() would 
never execute.