Issue 1659: (SegmentStore) BugFix for possible out-of order writes. (#1667)
Fixes a bug in SegmentStore where it would be possible to write two or more events received in a 
short timespan, in a different order than received from the caller. The actual changes are:

* StreamSegmentService: refactored all calls so they synchronously invoke the underlying 
Container's code vs via a CompletableFuture callback.
* StreamSegmentMapper: properly queueing up pending requests in the correct order (in which 
they were received) and then invoking them in the same order. This makes the invocation order predictable and no longer keeps us at the mercy of the JVM scheduler.
* OperationLog/DurableLog: add() returns a CompletableFuture since nobody cared about the 
result anyway. This enabled the simplification of some calls, thus reducing the number of invoked 
callbacks.