Issue 1716: (SegmentStore) Reuse physical buffers when creating new DataFrames (#1719)
Previously, every time we create a `DataFrame` a new 1MB buffer is allocated, regardless of how big the frame ends up being (whether 1KB or 1MB). This 1MB buffer is pointed to as long as the `DataFrame` exists, which is only after it is written to BookKeeper. In case BookKeeper is backed up, a lot of these frames can exist at any given times, hence a lot of 1MB buffers.

In this change, we attempt to improve the situation by reusing the previously used buffer, if it still has capacity. Main points:
- We keep track of the length of the last 10 built `DataFrames`.
- DataFrames already take a `ByteArraySegment` as input. The `DataFrameOutputStream` has been changed to return a sub-segment of a physical byte array vs a `ByteArraySegment` that wraps a new byte array.
- We also keep track of the last used offset inside this physical array. When asked to allocate a new `DataFrame` it tries to use the same buffer, but from the last used offset (until its end).
- This buffer is also released when: 
    - Not enough space is available (based on the average of the last 10 `DataFrame` lengths).
    - We have no more data to write (no more operations, thus not a high chance of requiring a `DataFrame` soon), thus releasing the 1MB buffer.
    - Since the buffer is used by consecutive `DataFrame`s, it will be released to the GC rather quickly, as soon as the last `DataFrame` that uses it is written to BK.
- With this change, we only use as much memory as we actually need, with very little waste (much less than before).