Issue 1718: Fix failing unit tests in Jenkins (#1746)
Test 1: StreamSegmentMapperTests.testGetOrAssignStreamSegmentIdWithMetadataLimit

* Fix merged with PR #1732 did not address the underlying issue, which was that subsequent 
requests would piggyback on the first one and never actually invoke the DurableLog.add() method 
which would've thrown the exception (and thus recorded that it happened).

* Fixes the issue by using different segment names, so we don't end up piggybacking on previous 
segments' callbacks.

Test 2: ReadIndex.triggerFutureReads

* This was not due to the unit test, rather it looks like a bug in the code with the original 
implementation of a fix for a unit test in the past. The conditions were reversed (compared to the 
description).