Issue 1760: Fixing HDFSStorage for small clusters (#1781)
StreamSegmentContainer:
* Closes Storage when closing the StreamSegmentContainer object itself. Closing Storage also 
closes the DFSClient, which closes any open streams (if any), thus releasing resources.

HDFSStorage:
* Disables the FS Cache. Each instance of HDFSStorage has exactly one instance of a 
DistributedFileSystem class (vs using a shared one across the process). This enables us to properly 
close that instance when the SegmentContainer shuts down, thus releasing resources.
* Adds the option to disable data node replacement if needed. According to the documentation, 
this is recommended for very small clusters to prevent the exact issues we seem to be dealing with.

FileSystemOperation/createFile: 
* Renames createEmptyFile to atomicCreate and made it atomic. HDFS file creation is a multi-step 
process (which includes creating the file and then closing the returned stream), which can result in 
exceptions when attempting to create the same file concurrently by multiple callers (none of which 
is FileAlreadyExistsException/StreamSegmentExistsException). Since a SegmentContainer is the 
only one creating Segments, and there is only one instance of a particular Container running at any 
given time, this was mitigated by using a simple pendingFileCreation set that keeps track of 
pending create requests.