Issue 1865: (BugFix) BookKeeperLog can stop processing writes in presence of an unhandled exception (#1866)
Fixes a bug in BookKeeperLog that can cause it to stall the Write Processor if an unhandled 
exception bubbles up to the main method.

* Both the WriteProcessor and RolloverProcessor use the SequentialAsyncProcessor to control their 
behavior. If either of their main methods allows an exception bubble out of them, it would be 
possible for them not to be reinvoked again, even if there was work left to do.

* SequentialAsyncProcessor has been fixed to catch all exceptions, run them through an exception 
handler and re-run the task if the handler indicates so.

* BookKeeperLog auto-closes if either the WriteProcessor or RolloverProcessor fails to execute 
more than a number of times in a row (based on an existing config value).