Issue #2738: Optimize re-filling read buffer on the client. (#2743)
1. Allow suggested read size to grow in response to available buffer capacity.
2. Issue reads calls to the server earlier.

Allows the suggested size grow to the remaining capacity of the buffer (This should avoid having to issue multiple requests to fill an empty buffer). And allows requests to be issued even when there is not yet capacity for the result to be moved into the buffer. (The class already handles this case and has tests for it.) This allows the request to go over the wire while the client is still working on data before it.

Signed-off-by: Andrei Paduroiu <andrei.paduroiu@emc.com>