Issue 2556: Expand client batch size to halve server latency (#2667)
When the server is under load, latency for the client increases rapidly. Depending on the event size this may cause the server to be CPU bound. This change automatically adjusts the batch size based on the server latency. This ensures the server has less work to perform when it is under load. This should lower append latency on moderately heavily loaded servers.

Signed-off-by: Andrei Paduroiu<andrei.paduroiu@emc.com>