Issue 2782: Validate read and write count after the test
Description:
Currently, we are not validating the write and read count after the test explicitly
This patch adds counters for write and read events and verifies
whether they are equal at the end of tests.

Purpose of code:
Ensure the reader and writer are matching in number

Code changes:
Added variables - readCount and writeCount, incremented
read event count if the read is successful and write events is set to NUM_EVENTS
(as there is no status return for writeEvent() function)
and validate their matching after the read is completed.

Signed-off-by: Ajesh JS <ajesh.johnsuseelabai@emc.com>