Issue 4751: ListScopes, CheckStreamExists and CheckScopeExists (#4950)
Implementation
Signed-off-by: Shivesh Ranjan <shivesh.ranjan@gmail.com>
Co-authored-by: Andrei Paduroiu <andreipaduroiu@users.noreply.github.com>

Change log description
Implementation for listScopes, CheckScopeExists and CheckStreamExists

Purpose of the change
Fixes #4751

What the code does
Adds grpc implementation for listScopes which is a paginated call.
Also provides implementation for CheckScopeExists and CheckStreamExists.
There is not implementation for Zookeeper based store for listScopes

How to verify it
Unit test updated