Issue 4956: Fix bugs in list streams and list scope api when resources are filtered (#4957)
Implementation
Signed-off-by: Shivesh Ranjan <shivesh.ranjan@gmail.com>

Change log description
If resources are filtered due to authorization, keep calling list resource api on store until we have collected desired number of results to fill the page or entries are exhausted.

Purpose of the change
Fixes #4956

What the code does
Make recursive call if the resources are filtered out due to authorization until we either collect "limit" number of records or exhaust all entries in the store.
This is performed for both streams and KVT.
KVT already had the implementation but there were bugs in it which are also fixed here.

Also, segment store returns an empty continuation token if we request from the last element and there are no more elements to read. This is handled in PravegaTableStoreHelper to return the caller's token if response from segment store is empty.

How to verify it
unit tests added

Signed-off-by: Ravi Rajamani <ravi.rajamani@emc.com>