Issue 4975: Replace Loading Cache with Regular Cache (#4992)
Signed-off-by: Shivesh Ranjan shivesh.ranjan@gmail.com

Change log description
Replace the cache implementation to not load the completablefuture into the cache as that can cause consistency issues.

Purpose of the change
Fixes #4975

What the code does
Changes the cache to load the value. The store first retrieves the value for mutable records from the cache. and if not found, it calls the store and loads the value from the future.
We dont try to optimize on multiple concurrent invocations to the loading function because that will result in the original problem of consistency check failure.

How to verify it
All existing tests should continue to pass.