Issue 5119: (SegmentStore) Copy-on-Read for Table Segment Compaction (#5120)
Updated the ReadResult interface with the ability to specify "copy-on-read" behavior for any data that is returned from the cache. If set, any data that is returned from the cache is copied into a heap buffer and served from there. Default is not set.

Updated TableCompactor to make use of this functionality.

Signed-off-by: Andrei Paduroiu <andrei.paduroiu@emc.com>

Co-authored-by: Sachin Jayant Joshi <44757683+sachin-j-joshi@users.noreply.github.com>
Signed-off-by: ManishKumarKeshri <manish.keshri562@gmail.com>