Issue 5181: Fix Null Pointer Exception in the log message (#5182)
Signed-off-by: Shivesh Ranjan shivesh.ranjan@gmail.com

Change log description
Fixes NPE in StreamMetadataTasks.retention method where we may truncate at an older streamcut despite not producing a "new stream cut".

Purpose of the change
Fixes #5181

What the code does
NPE is caused by a new log message that was introduced recently which logs newRecord instead of truncationRecord. The new record represents the new streamcut which is generated periodically. It is possible that we do not generate a new streamcut in each cycle if the previous cycle was cut short for some reason and a streamcut was generated. The truncation record represents an older streamcut at which we want to truncate.
The intent of the log message is to log the identified truncation streamcut. Instead it ends up logging new streamcut which can be null in some cases.

How to verify it
All tests should pass and there should be no NPE.