Issue 5111: Consumption based retention policy implementation (#5289)
Signed-off-by: Shivesh Ranjan <shivesh.ranjan@gmail.com>

## Change log description
Implements consumption based retention policy.

This does two things primarily -
In Stream Metadata Tasks it consolidates and computes lower bound on subscriber streamcuts and truncates at the said stream cut. This also applies time/size bounds
In stream cut validation in stream metadata store, we assumed stream cuts to be made of non overlapping segments, but a stream cut could have future segments which may not have been read from. These are represented using "-1" as the offset. So we update the logic in stream metadata store for validation as well as size computations for streamcuts.

Signed-off-by: Tom Kaitchuck <tom.kaitchuck@emc.com>