Changed all segment attributes from UUIDs to AttributeIds.
Defined AttributeId which is the base Attribute Id abstraction. Specific implementation is AttributeId.UUID (which is essentially all Core attributes and Streaming Segment Attributes - backward compatible). New implementation is AttributeId.Variable which will be used for any attribute ids that are not 16 bytes.

AttributeId generalizes the attribute structure; there is no longer a MSB or LSB; there are bit groups - this will make interoperability with AttributeId.Variable simpler.

Added/Updated StreamSegmentAppendOperation2 and UpdateAttributesOperation to support variable attributes.

Lots of places in the code had to be updated to use AttributeId instead of UUID - but otherwise the behavior is identical.

Signed-off-by: Andrei Paduroiu <andrei.paduroiu@emc.com>