Issue 5938: (KVT Beta 2) New Client API and Base Implementation. (#5960)
Updated KeyValueTable.java to reflect the new API proposed in PDP-48.
Updated KeyValueTableImpl.java to implement new contracts (except Iterators)
Updated TableKey.java and TableEntry.java to support new API contract.

NOTE: this is not a complete change. This PR will break KVT functionality, which is why several tests have been commented out and/or ignored (each has a TODO pointing to a GitHub issue to resolve it and the feature-kvt-beta2 branch should not be merged to master before those are done). Those items will be resolved in future PRs against this branch.

Signed-off-by: Andrei Paduroiu <andrei.paduroiu@emc.com>