Issue 6145: Custom thread pool (#6134)
This provides a custom impementation of the same interfaces as ScheduledThreadPoolExecutor in the standard library. However instead of using a heap data structure under lock to store the queued messages, it uses a lock-free queue. The queue works by splitting the delayed and non-delayed tasks. The non-delayed tasks are in a dequeue so adding and removing them can be done in O(1) without holding a lock. The delayed tasks still take O(log(n)) but do so with a lock-free data structure so that we can have parallelism when adding and removing them.

Signed-off-by: Tom Kaitchuck <tom.kaitchuck@emc.com>