Issue 6269: (Bugfix) Controller failed to update readergroup (#6270)
Signed-off-by: Shashwat Sharma <shashwat_sharma@dell.com>
Fixes #6270 
#Change log description
In ZKCheckpointStore, the AtomicBoolean isZKConnected is initialized to {{false. }}In the constructor, we add a listener that listens for state changes to the Curator Client connection state. However, by the time this listener is added, the ZKConnectionState is already "CONNECTED" and there are no further state changes, hence the value of isZKConnected remains false and never changes to true though the curator client is connected to Zookeeper. This causes ZKCheckpointStore.isZKConnected() to always return false.