Issue 6523: ContainerEventProcessor internal segment continuous truncations prevent StorageWriter batching (#6527) (#6584)
Adds the possibility to configure the truncation data side to ContainerEventProcessor. Set to 128KB the default truncation size for SLTS GC. Also, refresh the internal DirectSegmentAccess on each processing iteration of ContainerEventProcessor. Increased ContainerEventProcessor timeout from 100ms to 1 second, to reduce log noise when there are no items to process.

Signed-off-by: Raúl Gracia <raul.gracia@emc.com>