Issue 6616: Pinned Segments my not be pinned after a container restart.
Issue 6616: Pinned Segments my not be pinned after a container restart

Fixed a bug in how Segments are pinned via MetadataStore::registerPinnedSegment that was directly accessing the in-memory metadata without going through the underlying implementation of metadata store (e.g., Table Segments). Also, refactored the way the internal Segment for ContainerEventProcessor is created and pinned to memory.

Signed-off-by: Raúl Gracia <raul.gracia@emc.com>